src/models/wavelet_attention.py
"""
Multi-Scale Wavelet Attention with MODWT (Section 3.8)
Implements Equations 53-66 with:
- Maximal Overlap Discrete Wavelet Transform
- Regime-dependent scale attention
- Multi-head aggregation
"""

import torch
import torch.nn as nn
import torch.nn.functional as F
import pywt
import numpy as np
from typing import Optional, Dict, List


class MultiScaleWaveletAttention(nn.Module):
    """
    MODWT + Attention for multi-scale temporal analysis
    
    Paper Equations 53-66:
    - MODWT decomposition into multiple scales
    - Regime-dependent attention over scales
    - Multi-head fusion
    
    Scales:
    - Level 1: 2-4 days (ultra-short-term)
    - Level 2: 4-8 days (short-term)
    - Level 3: 8-16 days (medium-term)
    - Level 4: 16-32 days (monthly)
    - Level 5: 32-64 days (quarterly)
    - Smooth: >64 days (long-term trend)
    """
    
    def __init__(
        self,
        input_dim: int,
        wavelet_type: str = 'db8',
        decomposition_levels: int = 5,
        attention_heads: int = 4,
        regime_dependent: bool = True,
        num_regimes: int = 4,
        scale_specific_projection: bool = True
    ):
        super().__init__()
        self.input_dim = input_dim
        self.wavelet_type = wavelet_type
        self.decomposition_levels = decomposition_levels
        self.attention_heads = attention_heads
        self.regime_dependent = regime_dependent
        self.num_regimes = num_regimes
        self.scale_specific_projection = scale_specific_projection
        
        # Wavelet filters (precompute)
        wavelet = pywt.Wavelet(wavelet_type)
        self.register_buffer('wavelet_dec_lo', torch.tensor(wavelet.dec_lo, dtype=torch.float32))
        self.register_buffer('wavelet_dec_hi', torch.tensor(wavelet.dec_hi, dtype=torch.float32))
        
        # Number of scales = decomposition_levels + 1 (approximation + details)
        num_scales = decomposition_levels + 1
        
        # === Regime-dependent scale attention (Equations 58-59) ===
        if regime_dependent:
            self.scale_attention = nn.ModuleDict({
                f'regime_{k}': nn.Sequential(
                    nn.Linear(input_dim + num_regimes, 128),
                    nn.ReLU(),
                    nn.Dropout(0.1),
                    nn.Linear(128, num_scales)
                ) for k in range(num_regimes)
            })
        else:
            self.scale_attention = nn.Sequential(
                nn.Linear(input_dim, 128),
                nn.ReLU(),
                nn.Linear(128, num_scales)
            )
        
        # === Scale-specific projection (Equation 61) ===
        if scale_specific_projection:
            self.scale_projections = nn.ModuleList([
                nn.Linear(input_dim, input_dim) for _ in range(num_scales)
            ])
        
        # === Multi-head attention for scale aggregation (Equations 63-65) ===
        self.multihead_attn = nn.MultiheadAttention(
            embed_dim=input_dim,
            num_heads=attention_heads,
            dropout=0.1,
            batch_first=True
        )
        
        self.norm = nn.LayerNorm(input_dim)
        
    def forward(
        self,
        x: torch.Tensor,
        regime_labels: Optional[torch.Tensor] = None
    ) -> torch.Tensor:
        """
        Forward pass with MODWT decomposition and attention
        
        Args:
            x: (batch, input_dim) embeddings
            regime_labels: (batch,) regime assignments
            
        Returns:
            output: (batch, input_dim) multi-scale aggregated representation
        """
        batch_size = x.size(0)
        device = x.device
        
        # === STEP 1: MODWT Decomposition (Equations 54-57) ===
        wavelet_components = self._modwt_decomposition(x)  # (batch, num_scales, input_dim)
        num_scales = wavelet_components.size(1)
        
        # === STEP 2: Regime-Dependent Scale Attention (Equations 58-60) ===
        if self.regime_dependent and regime_labels is not None:
            attention_weights = []
            
            for i in range(batch_size):
                regime_id = regime_labels[i].item()
                regime_onehot = F.one_hot(regime_labels[i], self.num_regimes).float()
                
                # Concatenate embedding with regime
                attn_input = torch.cat([x[i], regime_onehot], dim=-1)
                
                # Compute attention weights for this regime (Equation 58)
                attn_w = F.softmax(
                    self.scale_attention[f'regime_{regime_id}'](attn_input),
                    dim=-1
                )
                attention_weights.append(attn_w)
            
            attention_weights = torch.stack(attention_weights, dim=0)  # (batch, num_scales)
        else:
            # Uniform or learned attention without regime conditioning
            attn_logits = self.scale_attention(x) if not self.regime_dependent else self.scale_attention['regime_0'](torch.cat([x, torch.zeros(batch_size, self.num_regimes, device=device)], dim=-1))
            attention_weights = F.softmax(attn_logits, dim=-1)  # (batch, num_scales)
        
        # === STEP 3: Scale-Specific Projection (Equation 61) ===
        if self.scale_specific_projection:
            projected_components = []
            for s in range(num_scales):
                component_s = wavelet_components[:, s, :]  # (batch, input_dim)
                projected_s = self.scale_projections[s](component_s)
                projected_components.append(projected_s)
            wavelet_components = torch.stack(projected_components, dim=1)  # (batch, num_scales, input_dim)
        
        # === STEP 4: Weighted Aggregation (Equation 62) ===
        aggregated = torch.sum(
            wavelet_components * attention_weights.unsqueeze(-1),
            dim=1
        )  # (batch, input_dim)
        
        # === STEP 5: Multi-Head Attention Refinement (Equations 63-65) ===
        # Treat scales as sequence for attention
        refined, attn_map = self.multihead_attn(
            wavelet_components,
            wavelet_components,
            wavelet_components
        )  # (batch, num_scales, input_dim)
        
        # Weighted aggregation of refined features
        refined_aggregated = torch.sum(
            refined * attention_weights.unsqueeze(-1),
            dim=1
        )  # (batch, input_dim)
        
        # === STEP 6: Residual Connection (Equation 66) ===
        output = self.norm(x + refined_aggregated)
        
        return output
    
    def _modwt_decomposition(self, x: torch.Tensor) -> torch.Tensor:
        """
        Apply MODWT decomposition to each feature dimension
        
        Args:
            x: (batch, input_dim) embeddings
            
        Returns:
            components: (batch, num_scales, input_dim) wavelet components
        """
        batch_size, input_dim = x.size()
        device = x.device
        num_scales = self.decomposition_levels + 1
        
        components = torch.zeros(batch_size, num_scales, input_dim, device=device)
        
        for i in range(batch_size):
            embedding = x[i].cpu().numpy()
            
            try:
                # MODWT decomposition using PyWavelets
                coeffs = pywt.wavedec(
                    embedding,
                    self.wavelet_type,
                    mode='periodization',
                    level=self.decomposition_levels
                )
                
                # coeffs = [cA_n, cD_n, cD_n-1, ..., cD_1]
                # cA_n: approximation (smooth, low-frequency)
                # cD_j: details at level j (high-frequency)
                
                # Pad all coefficients to same length (input_dim)
                for j, coeff in enumerate(coeffs):
                    if len(coeff) < input_dim:
                        # Pad with zeros
                        coeff_padded = np.pad(coeff, (0, input_dim - len(coeff)), mode='constant')
                    elif len(coeff) > input_dim:
                        # Truncate
                        coeff_padded = coeff[:input_dim]
                    else:
                        coeff_padded = coeff
                    
                    components[i, j] = torch.tensor(coeff_padded, device=device, dtype=torch.float32)
            
            except Exception as e:
                # Fallback: use original embedding for all scales
                print(f"Warning: MODWT decomposition failed for sample {i}: {e}")
                for j in range(num_scales):
                    components[i, j] = x[i]
        
        return components
    
    def get_scale_descriptions(self) -> List[str]:
        """Get human-readable descriptions of each scale"""
        descriptions = [
            "Long-term trend (>64 days)",
            "Quarterly cycle (32-64 days)",
            "Monthly cycle (16-32 days)",
            "Medium-term (8-16 days)",
            "Short-term (4-8 days)",
            "Ultra-short-term (2-4 days)"
        ]
        return descriptions[:self.decomposition_levels + 1]
    
    def visualize_scales(self, x: torch.Tensor) -> Dict:
        """Decompose and return all scale components for visualization"""
        with torch.no_grad():
            components = self._modwt_decomposition(x)
        
        return {
            f"scale_{i}": components[:, i, :].cpu().numpy()
            for i in range(components.size(1))
        }
