requirements.txt
torch==2.1.0
torchvision==0.16.0
pytorch-lightning==2.1.2
transformers==4.35.2
tokenizers==0.15.0
sentencepiece==0.1.99
accelerate==0.25.0
tsai==0.3.5
sktime==0.25.0
statsmodels==0.14.1
PyWavelets==1.4.1
pytorch-wavelets==1.3.0
torch-geometric==2.4.0
torch-scatter==2.1.2
torch-sparse==0.6.18
yfinance==0.2.32
pandas-datareader==0.10.0
alpha_vantage==2.3.1
finnhub-python==2.4.18
newsapi-python==0.2.7
ta==0.11.0
pandas_ta==0.3.14b0
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4
scikit-learn==1.3.2
river==0.19.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
shap==0.43.0
hmmlearn==0.3.0
optuna==3.4.0
wandb==0.16.1
tensorboard==2.15.1
tqdm==4.66.1
pyyaml==6.0.1
python-dotenv==1.0.0
pytest==7.4.3
requests==2.31.0
